pwb integration: corrupt.py
Further work has been done to integrate pywikibot. corrupt.py has had functions created using PWB to accomplish
the same tasks as mwclient. The only change needed for a complete switch of this file is either the renaming of
some function calls or removing the mwclient equivalents and getting rid of the "_PWB" function suffix in
image_corruption_utils.py.

Bug: #13