pwb integration: image_followup.py
image_followup.py has had functions created using PWB to accomplish the same tasks as mwclient. The only change
needed for a complete switch of this file is either the renaming of some function calls or the removal of the
mwclient equivalents and getting rid of the "_PWB" function suffix in image_followup.py and
image_corruption_utils.py.

Bug: #13