Add checks to ensure page run not redirect
Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125Add checks to ensure page run not redirect

Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125Add checks to ensure page run not redirect

Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125Add checks to ensure page run not redirect

Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125Add checks to ensure page run not redirect

Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125Add checks to ensure page run not redirect

Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125Add checks to ensure page run not redirect

Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125Add checks to ensure page run not redirect

Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125Add checks to ensure page run not redirect

Files run should not be redirects. If they are, spit out debug message and continue to the next file.

Closes T125